feat: introduce RouteProject Lambda as per-project dispatcher (#41)
